WIP - implements tuner source manager to calculate and update the tuner's  center frequency each time we allocate a new DDC/channel.  Adds  support for broadcasting channel count change source events so that the  tuner can accurately convey sourced channel counts via tuner events.
 Completes changeover of polyphase channel manager to support arbitrary
 bandwidth channels.  The new approach provides greater compatibility
 across any tuner sample rates.  Likewise, all decoders will have to be
 updated to adapt to any channel sample rate, although channel bandwidths
 will be at least 12.5 kHz and likely no more than 25.0 kHz, with a 2x
 oversampling that produces channel sample rates in range 25.0 - 50.0
 kHz.

+++ src/main/java/io/github/dsheirer/dsp/filter/channelizer/ChannelCalculator.java
@@ -185,7 +185,7 @@ public class ChannelCalculator
      */
     public double getChannelSampleRate()
     {
-        return (double)getChannelBandwidth() * getOversampling();
+        return getChannelBandwidth() * getOversampling();
     }
 
     /**
@@ -550,9 +550,9 @@ public class ChannelCalculator
     {
         mLog.debug("Starting ....");
 
-        int channelBandwidth = 12500;
-        int sampleRate = 100000; //Airspy
-        int channelCount = sampleRate / channelBandwidth;
+        double channelBandwidth = 12500.0;
+        double sampleRate = 100000.0; //Airspy
+        int channelCount = (int)(sampleRate / channelBandwidth);
         double centerFrequency = 455000000;
 
         ChannelCalculator calculator = new ChannelCalculator(sampleRate, channelCount, centerFrequency, 2.0);
